Refactoring #3
