+++ vividus-plugin-saucelabs/src/main/resources/properties/defaults/default.properties
@@ -1,0 +1,4 @@
+saucelabs.sauce-connect.enabled=false
+saucelabs.sauce-connect.use-latest-version=true
+saucelabs.sauce-connect.rest-url=
+saucelabs.sauce-connect.data-center=US
