+++ vividus-plugin-saucelabs/src/main/resources/spring.xml
@@ -7,16 +7,25 @@
     <bean id="sauceConnectManager" class="org.vividus.selenium.sauce.SauceConnectManager">
         <property name="sauceLabsUsername" value="${selenium.grid.username}"/>
         <property name="sauceLabsAccessKey" value="${selenium.grid.password}"/>
+        <property name="testContext" ref="testContext"/>
         <property name="sauceTunnelManager">
             <bean class="org.vividus.selenium.sauce.SynchronizedUnpackingSauceConnectTunnelManager">
                 <constructor-arg value="false"/>
                 <property name="useLatestSauceConnect" value="${saucelabs.sauce-connect.use-latest-version}"/>
+                <property name="sauceRest">
+                    <bean class="com.saucelabs.saucerest.SauceREST">
+                        <constructor-arg index="0" value="${selenium.grid.username}" />
+                        <constructor-arg index="1" value="${selenium.grid.password}" />
+                        <constructor-arg index="2" value="${saucelabs.sauce-connect.data-center}" />
+                    </bean>
+                </property>
             </bean>
         </property>
     </bean>
 
     <bean class="org.vividus.selenium.sauce.SauceLabsCapabilitiesConfigurer">
         <property name="sauceConnectEnabled" value="${saucelabs.sauce-connect.enabled}"/>
+        <property name="restUrl" value="${saucelabs.sauce-connect.rest-url}" />
     </bean>
 
 </beans>
