+++ vividus-tests/src/main/resources/properties/configuration.properties
@@ -1,3 +1,3 @@
-configuration.profiles=web/headless/chrome
+configuration.profiles=web/desktop/chrome
 configuration.suite=integration
 configuration.environments=
